pyproject.toml
[project]
name = "rebuild-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "asyncpg==0.31.0",
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain-openai>=1.0.1",
    "lxml>=6.0.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pydantic-settings>=2.13.1",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[[tool.uv.index]]
url = "https://pypi.org/simple/"
default = true
